+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Latest changed
 
+*   Fix rgb camera attributes
+*   Add intensity value to point cloud message
 *   Fixed wrong TF for ego_vehicle
 *   Improve version check
 *   Fix cleanup